Refactor PinMapping.
Preparing for cable management.
Asynchronous connections are not saved in lists, but in unordered_map associated with their row (if there is no raster, each device pin should be associated with own "row" internally).
For each row, the connected devices and pins are saved separately (therefore, pins are not directly associated with devices anymore).

Next steps:
- cleaning up these changes
- adding connections with dialog (should always be an option since dragging cables would not be possible without raster, e.g. oled_shield.json)
- using that dialog to update the other configs
- adding cables with drag and drop if Embedded is visible