Refactor and Cleanup.
reading_connections and writing_connections are back.
Connections are now held in the raster for basic information about what is where and in the old connection structures for fast access.
SPI function is always set for one pin (and therefore all connected devices).
Pin sync is set for each device separately. writing/reading is also for each device, as that originates from the pin directions saved in the PinLayout.
TODO:
- setSPI/Pin sync methods
- update other configs
- draw cables
- input